feat(profileHeader): make revalidate cache and make server component

+++ src/lib/actionService.ts
@@ -1,47 +1,56 @@
 'use server';
 
 import prisma from 'src/lib/prisma';
+import { revalidateTag } from 'next/cache';
+import { auth } from '@/auth';
 
-export async function follow (currentUserId: string, targetUserId: string) {
+export async function follow (targetUserId: string) {
+  const session = await auth();
+  if (!session) {
+    throw new Error('User is not authenticated.');
+  }
   try {
     await prisma.userRelationship.create({
       data: {
-        follower_user_id: currentUserId,
+        follower_user_id: session.user.id,
         followed_user_id: targetUserId,
       },
     });
-    return true;
   } catch (error) {
-    console.error('Error following user:', error);
-    return false;
+    throw new Error(`Error following user: ${error}`);
   }
+  revalidateTag('user');
+  // revalidatePath(`/dashboard/user/${targetUserId}`);
 }
 
-export async function unfollow (currentUserId: string, targetUserId: string) {
+export async function unfollow (targetUserId: string) {
   try {
+    const session = await auth();
+    if (!session) {
+      throw new Error('User is not authenticated.');
+    }
     await prisma.userRelationship.deleteMany({
       where: {
-        follower_user_id: currentUserId,
+        follower_user_id: session.user.id,
         followed_user_id: targetUserId,
       },
     });
-    return true;
   } catch (error) {
-    console.error('Error unfollowing user:', error);
-    return false;
+    throw new Error(`Error Unfollowing user: ${error}`);
   }
+  revalidateTag('user');
+  // revalidatePath('/dashboard/', 'page');
 }
 
 export async function handleActionButtonClick (
-  currentUserId: string,
-  profileUserId: string,
   actionButtonValue: string,
+  profileId: string,
 ) {
   if (actionButtonValue === 'Follow' || actionButtonValue === 'Follow Back') {
-    return follow(currentUserId as string, profileUserId as string);
+    return follow(profileId);
   }
   if (actionButtonValue === 'Unfollow') {
-    return unfollow(currentUserId as string, profileUserId as string);
+    return unfollow(profileId);
   }
   console.log('NOT IMPLEMENTED YET');
   return false;
